requirements.txt
Flask==2.3.2
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.0
werkzeug==2.3.6
